+++ backend/package.json
@@ -11,15 +11,13 @@
     "test": "echo \"No tests yet...\""
   },
   "dependencies": {
-    "@azure/data-tables": "^13.2.2",
-    "@azure/functions": "^4.0.0",
-    "install": "^0.13.0",
-    "npm": "^10.8.1"
+    "@azure/abort-controller": "^2.1.2",
+    "@azure/data-tables": "^13.2.2"
   },
   "devDependencies": {
-    "@types/node": "^20.x",
+    "@azure/functions": "^2.0.0",
+    "@types/node": "^14.x",
     "rimraf": "^5.0.0",
     "typescript": "^4.0.0"
-  },
-  "main": "dist/src/{index.js,functions/*.js}"
+  }
 }
