--- backend/src/index.ts
@@ -1,5 +1,0 @@
-import { app } from '@azure/functions';
-
-app.setup({
-    enableHttpStream: true,
-});
